[#11290] log4j-it, extract testbase
