+++ agent-module/plugins-it/log4j-it/src/test/java/com/navercorp/pinpoint/it/plugin/log4j/Log4jIT.java
@@ -17,16 +17,12 @@ package com.navercorp.pinpoint.it.plugin.log4j;
 
 import com.navercorp.pinpoint.it.plugin.utils.AgentPath;
 import com.navercorp.pinpoint.it.plugin.utils.PluginITConstants;
-import com.navercorp.pinpoint.it.plugin.utils.StdoutRecorder;
 import com.navercorp.pinpoint.test.plugin.Dependency;
 import com.navercorp.pinpoint.test.plugin.ImportPlugin;
 import com.navercorp.pinpoint.test.plugin.PinpointAgent;
 import com.navercorp.pinpoint.test.plugin.PinpointConfig;
 import com.navercorp.pinpoint.test.plugin.PluginForkedTest;
 import com.navercorp.pinpoint.test.plugin.TransformInclude;
-import org.apache.log4j.Logger;
-import org.apache.log4j.MDC;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 @PluginForkedTest
@@ -35,63 +31,15 @@ import org.junit.jupiter.api.Test;
 @ImportPlugin({"com.navercorp.pinpoint:pinpoint-log4j-plugin"})
 @PinpointConfig("pinpoint-spring-bean-test.config")
 @TransformInclude("org.apache.log4j.")
-public class Log4jIT {
-
-    private Logger logger;
+public class Log4jIT extends Log4jTestBase {
 
     @Test
     public void test() {
-        Logger logger = Logger.getLogger(getClass());
-        logger.error("maru");
-
-        checkVersion(logger);
-
-        Assertions.assertNotNull(MDC.get("PtxId"), "txId");
-        Assertions.assertNotNull(MDC.get("PspanId"), "spanId");
+        checkMDC();
     }
 
     @Test
     public void patternUpdate() {
-        final String msg = "pattern";
-
-
-        StdoutRecorder stdoutRecorder = new StdoutRecorder();
-        final String log = stdoutRecorder.record(new Runnable() {
-            @Override
-            public void run() {
-                logger = Logger.getLogger("patternUpdateLogback");
-                logger.error(msg);
-            }
-        });
-
-        System.out.println(log);
-        Assertions.assertNotNull(log, "log null");
-        Assertions.assertTrue(log.contains(msg), "contains msg");
-        Assertions.assertTrue(log.contains("TxId"), "contains TxId");
-
-        Assertions.assertNotNull(logger, "logger null");
-        checkVersion(logger);
+        checkPatternReplace();
     }
-
-    private void checkVersion(Logger logger) {
-        final String location = getLoggerJarLocation(logger);
-        Assertions.assertNotNull(location, "location null");
-        System.out.println("Log4j jar location:" + location);
-
-        final String testVersion = getTestVersion();
-        Assertions.assertTrue(location.contains("/" + testVersion + "/"), "test version is not " + getTestVersion());
-    }
-
-    private String getTestVersion() {
-        final String[] threadInfo = Thread.currentThread().getName()
-                .replace(getClass().getName(), "")
-                .replace(" Thread", "")
-                .replace(" ", "").replace("log4j-", "").split(":");
-        return threadInfo[0];
-    }
-
-    private String getLoggerJarLocation(Object object) {
-        return object.getClass().getProtectionDomain().getCodeSource().getLocation().toString();
-    }
-
 }
